Backend/requirements.txt
fastapi
uvicorn
python-dotenv
diffusers
transformers
torch
pillow
